Used Food instead of GameObject in stockpile area
